+++ cmd/cli/cmd/environment/create/create.go
@@ -2,6 +2,7 @@ package create
 
 import "github.com/spf13/cobra"
 
+// NewCmd returns a cobra.Command for Capact environment related operations.
 func NewCmd() *cobra.Command {
 	root := &cobra.Command{
 		Use:   "create",
