+++ cmd/cli/cmd/environment/delete/kind.go
@@ -5,7 +5,7 @@ import (
 	"github.com/spf13/cobra"
 )
 
-// NewKind returns a cobra.Command for deleting Kind environment.
+// NewKind returns a cobra.Command for deleting kind environment.
 func NewKind() *cobra.Command {
 	var opts delete.KindOptions
